Move retain_mut backport to own mod and conditionally compile it
