+++ Cargo.toml
@@ -21,6 +21,10 @@ serde_json = "1.0.75"
 serde-value = "0.7"
 async-trait = "0.1.9"
 
+[dependencies.rustversion]
+version = "1.0.7"
+optional = true
+
 [dependencies.simd-json]
 version = "0.4.14"
 optional = true
@@ -166,7 +170,7 @@ default_no_backend = [
 
 builder = ["utils"]
 cache = ["dashmap", "parking_lot"]
-collector = ["gateway", "model"]
+collector = ["gateway", "model", "rustversion"]
 client = ["http", "typemap_rev"]
 extras = []
 framework = ["client", "model", "utils"]
