+++ src/utils/mod.rs
@@ -3,6 +3,7 @@
 
 #[cfg(feature = "client")]
 mod argument_convert;
+pub(crate) mod backports;
 mod colour;
 #[cfg(feature = "cache")]
 mod content_safe;
